refactor(web): apply code-review simplifications
Independent quality pass — no behavior change:

- Delete the unused Sparkline component (zero call sites) and its doc refs.
- Trim useTheme to its consumed surface (preference + cycleTheme); drop the
  unread resolvedTheme/setTheme and the matchMedia subscription that only fed
  them — auto-mode follows the OS via CSS, not JS.
- Collapse the getEmbeddingConfig tRPC procedure to reuse getEmbeddingConfigInfo
  instead of re-mapping the config literal.
- Factor the repeated live/persisted page-count extraction into jobPageCounts
  (JobCard ×2, FailedJobCard); use progressPercent in Overview's job row.
- Extract one formatRelativeShort helper; the Overview and library-detail
  relative-time formatters become thin null-guarding wrappers over it.
- Fix a stale getVersionComposition docstring (described the removed size/
  depth/chunks fields; it returns only mimeTypes now).
