+++ src/web/client/hooks/useTheme.ts
@@ -17,11 +17,7 @@ import { useCallback, useSyncExternalStore } from "react";
 /** A user-selectable theme preference. "auto" follows the OS setting. */
 export type ThemePreference = "light" | "dark" | "auto";
 
-/** The theme actually in effect once "auto" is resolved against the OS. */
-export type ResolvedTheme = "light" | "dark";
-
 const STORAGE_KEY = "docs-mcp-theme";
-const DARK_QUERY = "(prefers-color-scheme: dark)";
 
 function readStoredPreference(): ThemePreference {
   if (typeof window === "undefined") return "auto";
@@ -65,38 +61,21 @@ function getPreferenceSnapshot(): ThemePreference {
   return currentPreference;
 }
 
-function getSystemThemeSnapshot(): ResolvedTheme {
-  if (typeof window === "undefined" || !window.matchMedia) return "light";
-  return window.matchMedia(DARK_QUERY).matches ? "dark" : "light";
-}
-
-function subscribeSystemTheme(listener: () => void): () => void {
-  if (typeof window === "undefined" || !window.matchMedia) return () => {};
-  const mql = window.matchMedia(DARK_QUERY);
-  mql.addEventListener("change", listener);
-  return () => mql.removeEventListener("change", listener);
-}
-
 export interface UseThemeResult {
   /** The stored preference: "light", "dark", or "auto". */
   preference: ThemePreference;
-  /** The theme actually applied right now (auto resolved to light/dark). */
-  resolvedTheme: ResolvedTheme;
-  /** Sets an explicit preference and persists it. */
-  setTheme: (preference: ThemePreference) => void;
   /** Cycles auto -> light -> dark -> auto, matching the topbar toggle button. */
   cycleTheme: () => void;
 }
 
 /**
- * Reads and updates the current theme preference, keeping `<html data-theme>`
- * and `localStorage` in sync with the returned state.
+ * Reads the current theme preference and cycles it, keeping `<html data-theme>`
+ * and `localStorage` in sync. Auto-mode follows the OS purely via CSS (the
+ * `data-theme` attribute is removed so the `@media` block governs), so no
+ * system-theme listener is needed here.
  */
 export function useTheme(): UseThemeResult {
   const preference = useSyncExternalStore(subscribePreference, getPreferenceSnapshot);
-  const systemTheme = useSyncExternalStore(subscribeSystemTheme, getSystemThemeSnapshot);
-
-  const setTheme = useCallback((next: ThemePreference) => setPreference(next), []);
 
   const cycleTheme = useCallback(() => {
     setPreference(
@@ -108,7 +87,5 @@ export function useTheme(): UseThemeResult {
     );
   }, []);
 
-  const resolvedTheme: ResolvedTheme = preference === "auto" ? systemTheme : preference;
-
-  return { preference, resolvedTheme, setTheme, cycleTheme };
+  return { preference, cycleTheme };
 }
